requirements.txt
streamlit
opencv-python
numpy
mediapipe
Pillow
